Added authtorization logic in controllers Fixed postman's collections

+++ Habr/Habr.WebAPI/Controllers/CommentController.cs
@@ -1,5 +1,4 @@
-﻿using Habr.BusinessLogic.Helpers;
-using Habr.BusinessLogic.Interfaces;
+﻿using Habr.BusinessLogic.Interfaces;
 using Habr.Common.DTO;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,6 +19,11 @@ namespace Habr.WebAPI.Controllers
         [HttpPost("comments")]
         public async Task<IActionResult> CreateCommentAsync([FromBody] CreateCommentDTO comment)
         {
+            if (!JwtHelper.IsJwtIdClaimValid(HttpContext.User.Claims, comment.UserId))
+            {
+                return Forbid();
+            }
+
             var createdComment = await _commentService.CreateCommentAsync(comment);
             return CreatedAtAction(nameof(GetCommentsAsync), new { commentId = createdComment.Id }, createdComment);
         }
@@ -34,6 +38,11 @@ namespace Habr.WebAPI.Controllers
         [HttpDelete("users/{userId:int}/comments/{commentId:int}")]
         public async Task<IActionResult> DeleteCommentAsync([FromRoute] int userId, [FromRoute] int commentId)
         {
+            if (!JwtHelper.IsJwtIdClaimValid(HttpContext.User.Claims, userId))
+            {
+                return Forbid();
+            }
+
             await _commentService.DeleteCommentAsync(commentId, userId);
             return Ok();
         }
